Improve seeds and take tasks

+++ db/seeds.rb
@@ -23,7 +23,7 @@ Rake::Task['taxonomy:delete_old_taxonomies'].invoke
 Rake::Task['taxonomy:populate_parents'].invoke
 
 all_needs_met_taxonomy = Taxonomy.where(name: "All Needs Met").first
-all_needs_met_taxonomy.destroy!
+all_needs_met_taxonomy.destroy! if all_needs_met_taxonomy.present?
 
 puts "Took #{(end_time - start_time)/60} minutes"
